SimpleMathService: Renamed all references from Service1 to AdditionOperations.  Added the ShouldLogSoapRequestsAndResponses appSettings key.  Moved all using statements inside the namespaces.
SoapRequestAndResponseTracing: Initial check-in missed DebugMessageDispatcher.cs; adding that to repo.  Created an ILogger interface and a stub method called Log based on the interface (not actually logging to a file as of yet).  Calling the Log method if the ShouldLogSoapRequestsAndResponses appSetting in the caller exists and is true.  Moved all using statements inside the namespaces.